requirements.txt
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAudio==0.2.11
anthropic>=0.25.0